pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "PCL"
description = 'A Pytorch implementation of Prototypical Contrastive Learning (PCL) according to https://arxiv.org/abs/2005.04966'
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
keywords = ['pytorch', 'pcl', 'contrastive learning', 'prototypical contrastive learning']
authors = [
  { name = "Zegang Cheng", email = "zc2309@nyu.edu" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: Implementation :: CPython",
]
dependencies = [
  "pandas",
  "Pillow",
  "pykeops",
  "torch>=1.13.*",
  "torchvision>=0.14.*",
  "tqdm",
  "scipy",
  "pandas",
  "numpy",
  "click",
  "matplotlib",
  "jupyter",
]
dynamic = ["version"]

[project.urls]
Documentation = "https://github.com/chengzegang/PCL#readme"
Issues = "https://github.com/chengzegang/PCL/issues"
Source = "https://github.com/chengzegang/PCL"

[tool.hatch.version]
path = "pcl/__about__.py"

[tool.hatch.envs.default]
dependencies = [
  "pytest",
  "pytest-cov",
]
[tool.hatch.envs.default.scripts]
cov = "pytest --cov-report=term-missing --cov-config=pyproject.toml --cov=pytorch_vlad --cov=tests {args}"
no-cov = "cov --no-cov {args}"

[[tool.hatch.envs.test.matrix]]
python = ["310", "311"]

[tool.coverage.run]
branch = true
parallel = true
omit = [
  "pcl/__about__.py",
]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = [
    "torch",
    "torch.*",
    "torchvision",
    "torchvision.*",
    "pykeops.*",
]
ignore_missing_imports = true